Add mobile preparation according to: https://v2.tauri.app/start/migrate/from-tauri-1/
