+++ src-tauri/Cargo.toml
@@ -1,3 +1,7 @@
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 [package]
 name = "tauri_template"
 version = "0.0.0"
